Adds basic train tracking functionality
This makes a request on page load for the names of all light and heavy
rail lines on the mbta, then pushes their names into the view. Also
makes a call to each route and destermines how many trains are headed
what direction and also what those directions are named.