Adds Alert feeds
Alerts are now pulled in for each line and return the alert summary as
well as a link to read more about the alert if a user chooses to. Lines
now start collapsed and can be expanded by the user. Next steps are to
verrify that the alerts are current and add maps or another additional
feature.